Move classes into seperate file
